+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - feat: add `NesIcons.remove`
 - feat: add `NesIcons.redo`
 - feat: add `customExtensions`
-- feat: add keyboard support 
+- feat: add keyboard support
+- feat: add `backgroundColor` to `NesContainer`
 
 # 0.1.0